Move MainActivity to root of hydra/, remove ComingSoonFragment: it's better to fail fast with unkown ID's
