--- app/src/main/java/be/ugent/zeus/hydra/common/main/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Contains code for displaying the main activity in the app. This is the code that will tie all different functions
- * together.
- *
- * @author Niko Strijbol
- */
-package be.ugent.zeus.hydra.common.main;
