--- app/src/main/res/layout/fragment_comingsoon.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<android.support.v4.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_gravity="fill_vertical">
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="@string/comingsoontitle"
-            android:id="@+id/comingsoon"
-            android:textSize="48sp"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginTop="41dp" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="@string/openFacebook"
-            android:id="@+id/textView"
-            android:layout_gravity="center_horizontal"
-            android:gravity="center_horizontal"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp" />
-
-        <ImageButton
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            android:scaleType="fitCenter"
-            android:id="@+id/fbButton"
-            android:layout_marginTop="48dp"
-            android:layout_gravity="center_horizontal"
-            app:srcCompat="@drawable/ic_social_facebook"
-            android:tint="@color/bright_blue"
-            android:adjustViewBounds="true"
-            android:background="#00ffffff"
-            android:clickable="false"
-            android:contentDescription="@string/facebookButtonDescription" />
-
-    </LinearLayout>
-
-
-</android.support.v4.widget.NestedScrollView>
