+++ app/src/main/res/xml/shortcuts.xml
@@ -13,7 +13,7 @@
         <intent
             android:action="android.intent.action.VIEW"
             android:targetPackage="be.ugent.zeus.hydra"
-            android:targetClass="be.ugent.zeus.hydra.common.main.MainActivity">
+            android:targetClass="be.ugent.zeus.hydra.MainActivity">
             <extra android:name="argTabShortcut" android:value="1" />
         </intent>
     </shortcut>
@@ -29,7 +29,7 @@
         <intent
             android:action="android.intent.action.VIEW"
             android:targetPackage="be.ugent.zeus.hydra"
-            android:targetClass="be.ugent.zeus.hydra.common.main.MainActivity">
+            android:targetClass="be.ugent.zeus.hydra.MainActivity">
             <extra android:name="argTabShortcut" android:value="2" />
         </intent>
     </shortcut>
